[refact] 메인페이지 리팩토링
