+++ src/components/feed/DetailFeed.tsx
@@ -11,8 +11,8 @@ import LinkIcon from '../../assets/LinkWhite.svg'
 import MoreDots from '../../assets/MoreDots.svg'
 import { colors } from '../../styles/colors'
 import { BottomSheet } from 'react-spring-bottom-sheet'
-import pencil from '../../assets/feed/Pencil.svg'
-import trash from '../../assets/feed/Trash.svg'
+import pencil from '../../assets/main/Pencil.svg'
+import trash from '../../assets/main/Trash.svg'
 import Download from '../../assets/Download.svg'
 import TelePathyMotion from './TelepathyMotion'
