+++ src/components/main/Flips.tsx
@@ -1,7 +1,8 @@
 import styled from 'styled-components'
-import NoFeed from '../../assets/main/NoneFeed.svg'
-import { colors } from '../../styles/colors'
+import NoFeed from '@/assets/main/NoneFeed.svg'
+import { colors } from '@/styles/colors'
 
+// 해당 그룹에 피드가 없을 때 컴포넌트
 const Flips = () => {
   return (
     <Container>
@@ -19,9 +20,9 @@ export default Flips
 const Container = styled.div`
   display: flex;
   flex-direction: column;
-  width: 100%;
   flex-direction: center;
   align-items: center;
+  width: 100%;
   margin: 60px 0px;
   gap: 30px;
 `
@@ -37,24 +38,22 @@ const WarnText = styled.div`
   color: ${colors.grey3};
   font-family: Pretendard;
   font-size: 14px;
-  font-style: normal;
   font-weight: 600;
   line-height: 21px;
   letter-spacing: -0.56px;
 `
 const QBtn = styled.button`
   display: flex;
-  padding: 8px 24px;
   justify-content: center;
   align-items: center;
+  padding: 8px 24px;
   border-radius: 8px;
   border: none;
   outline: none;
-  background: ${colors.grey1};
+  background-color: ${colors.grey1};
   color: ${colors.white};
   font-family: Pretendard;
   font-size: 12px;
-  font-style: normal;
   font-weight: 600;
   line-height: 18px;
   letter-spacing: -0.24px;
