+++ src/components/main/LoginModal.tsx
@@ -1,17 +1,15 @@
-import { SearchModalBox, SearchModalContent } from '../common/ModalStyle'
 import styled from 'styled-components'
-import { colors } from '../../styles/colors'
-import KakaoIcon from '../../assets/KakaoIcon.svg'
 import { AnimatePresence } from 'framer-motion'
+import { SearchModalBox, SearchModalContent } from '@/components/common/ModalStyle'
+import { ModalProps } from '@/components/main/types'
+import { colors } from '@/styles/colors'
+import KakaoIcon from '@/assets/login/KakaoIcon.svg'
 
-interface ModalProps {
-  content: string
-  clickModal: () => void
-}
 const LoginModal = (props: ModalProps) => {
   // 전달받은 state 함수
   const { content, clickModal } = props
 
+  // 카카오 로그인
   const clientId = import.meta.env.VITE_KAKAO_CLIENT_ID
   const redirectUri = import.meta.env.VITE_KAKAO_REDIRECT_URI
   const link = `http://kauth.kakao.com/oauth/authorize?client_id=${clientId}&redirect_uri=${redirectUri}&response_type=code`
@@ -51,41 +49,38 @@ const LoginModal = (props: ModalProps) => {
 export default LoginModal
 
 const ModalContent = styled.div`
-  position: relative;
   display: flex;
   flex-direction: column;
-  width: 100%;
-  height: 100%;
   justify-content: center;
   align-items: center;
+  position: relative;
+  width: 100%;
+  height: 100%;
   gap: 8px;
 `
 const Content = styled.div`
+  margin: 20px 0px;
   color: ${colors.grey1};
   text-align: center;
   font-family: Pretendard;
-  font-size: 16px;
-  font-style: normal;
   font-weight: 600;
   line-height: 28.8px;
   letter-spacing: -0.32px;
-  margin: 20px 0px;
   white-space: pre-wrap;
 `
 const Button = styled.button<{ $positive: boolean }>`
   display: flex;
-  height: 56px;
-  padding: 16px 0px;
   justify-content: center;
   align-items: center;
   align-self: stretch;
+  height: 56px;
+  padding: 16px 0px;
   border-radius: 12px;
-  background: ${(props) => (props.$positive ? '#FEE500' : colors.white)};
+  background-color: ${(props) => (props.$positive ? '#FEE500' : colors.white)};
   border: 1px solid ${(props) => (props.$positive ? '#FEE500' : colors.grey5)};
   color: ${(props) => (props.$positive ? colors.grey1 : colors.grey3)};
   gap: 10px;
   font-size: ${(props) => (props.$positive ? '14px' : '16px')};
-  font-style: normal;
   font-weight: 600;
   line-height: ${(props) => (props.$positive ? '21px' : '24px')};
   letter-spacing: ${(props) => (props.$positive ? '-0.28px' : '-0.32px')};
