+++ src/components/main/MainProfile.tsx
@@ -1,9 +1,9 @@
 import styled from 'styled-components'
-import { colors } from '../../styles/colors'
-import DefaultImage from '../../assets/main/DefaultImage.svg'
 import { useEffect } from 'react'
 import { useRecoilState } from 'recoil'
-import { UserInfoStateProps, userInfoState } from '../../context/Atoms'
+import { colors } from '@/styles/colors'
+import { UserInfoStateProps, userInfoState } from '@/context/Atoms'
+import DefaultImage from '@/assets/main/DefaultImage.svg'
 
 declare global {
   interface Window {
@@ -15,6 +15,7 @@ interface MainProfileProps {
   // image: string 프로필 이미지도 받아오기
 }
 
+// 메인프로필 컴포넌트
 const MainProfile = ({ nickname }: MainProfileProps) => {
   const [userInfo, setUserInfo] = useRecoilState<UserInfoStateProps>(userInfoState)
 
@@ -74,7 +75,7 @@ const MainProfile = ({ nickname }: MainProfileProps) => {
         <Nickname>{nickname}</Nickname>
         <ShareButton onClick={sharing}>내 플리빗 초대</ShareButton>
       </ProfileContents>
-      {userInfo.profileImage === null ? (
+      {userInfo.profileImage === '' ? (
         <ProfileImage src={DefaultImage} />
       ) : (
         <ProfileImage src={userInfo.profileImage} />
@@ -107,23 +108,21 @@ const Nickname = styled.div`
   color: ${colors.grey1};
   font-family: Pretendard;
   font-size: 16px;
-  font-style: normal;
   font-weight: 700;
   line-height: 24px;
   letter-spacing: -0.32px;
 `
 const ShareButton = styled.button`
   display: flex;
-  padding: 8px 24px;
   justify-content: center;
   align-items: center;
+  padding: 8px 24px;
   border-radius: 8px;
   border: 0;
-  background: ${colors.grey7};
+  background-color: ${colors.grey7};
   color: ${colors.grey3};
   font-family: Pretendard;
   font-size: 12px;
-  font-style: normal;
   font-weight: 600;
   line-height: 18px;
   letter-spacing: -0.24px;
