--- src/components/main/types.ts
@@ -1,4 +1,0 @@
-export interface userDataProps {
-  nickname: string
-  memberId: number
-}
